requirements.txt
streamlit
langchain
langchain-community
fpdf
python-docx
huggingface_hub